requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
boto3==1.17.73
botocore==1.20.73
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==5.0.7
dj-database-url==0.5.0
Django==3.2.2
django-audiofield==0.10.0
django-celery==3.1.17
django-cors-headers==3.7.0
django-extensions==3.1.3
django-heroku==0.3.1
django-money==1.3.1
django-shopping-cart==0.1
django-storages==1.11.1
djangorestframework==3.12.4
forex-python==1.5
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
jsonpickle==2.0.0
kombu==5.0.2
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
py-moneyed==0.8.0
python-dateutil==2.8.1
python-decouple==3.4
python-paystack==1.2.0
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
validators==0.18.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
